modules/ambient/requirements.txt
azure-iot-device==2.7.1
Adafruit-DHT==1.4.0
